refactor(frontend): decompose AdminMonitoringPage into focused modules
Split 1,966-line monolith into 12 focused files under pages/admin/monitoring/:
- types.ts: all interfaces (~170 lines)
- constants.ts: JUSTICE_KINDS, DOC_FORMS, MONTHS_SHORT
- shared.tsx: formatDate, formatNumber, SectionLoader, SectionError, DataTable, SummaryCard, toTableInfoArray, completeness helpers
- CourtDocsSection.tsx: court documents by practice area with expandable rows
- BackfillProgress.tsx: backfill job progress display
- ScraperJobCard.tsx: individual scraper job card
- CourtDataMapSection.tsx: coverage heatmap + batch scraper management
- DocumentCompletenessSection.tsx: completeness check + backfill workflow
- DocumentModal.tsx: document viewer modal
- ImportSamplesSection.tsx: recent import samples with source expansion
- index.tsx: thin orchestrator (main page component)

Original file reduced to 3-line re-export shim for backward compatibility.
