+++ lexwebapp/src/pages/AdminMonitoringPage.tsx
@@ -1,1966 +1,4 @@
 /**
- * Admin Monitoring Page
- * Dashboard showing volume, source, and update frequency of all data sources
- * Loads each section independently for progressive rendering
+ * Re-export from admin/monitoring/ directory for backward compatibility.
  */
-
-import React, { useEffect, useState, useCallback } from 'react';
-import {
-  Activity,
-  Database,
-  RefreshCw,
-  CheckCircle,
-  AlertTriangle,
-  XCircle,
-  Server,
-  HardDrive,
-  ExternalLink,
-  Clock,
-  Layers,
-  Scale,
-  FileText,
-  ChevronDown,
-  ChevronUp,
-  Play,
-  Download,
-  Square,
-  Eye,
-  X,
-  Plus,
-  Grid3X3,
-} from 'lucide-react';
-import { api } from '../utils/api-client';
-
-interface TableInfo {
-  id: string;
-  name: string;
-  rows: number;
-  source: string;
-  sourceUrl: string;
-  updateFrequency: string;
-  lastUpdate: string | null;
-  lastBatchCount?: number;
-}
-
-interface CourtDocInfo {
-  id: string;
-  title: string;
-  date: string | null;
-  court: string | null;
-  case_number: string | null;
-  dispute_category: string | null;
-  loaded_at: string;
-}
-
-interface CourtCategory {
-  code: string;
-  name: string;
-  total: number;
-  recent: number;
-  earliest_date: string | null;
-  latest_date: string | null;
-  last_loaded_at: string | null;
-  documents: CourtDocInfo[];
-}
-
-interface CourtDocsData {
-  total_court_docs: number;
-  recent_court_docs: number;
-  days: number;
-  categories: CourtCategory[];
-}
-
-interface CompletenessResult {
-  checked_at: string;
-  runs_today: number;
-  max_runs_per_day: number;
-  summary: {
-    total_documents: number;
-    with_plaintext: number;
-    with_html: number;
-    with_only_html: number;
-    with_both: number;
-    missing_both: number;
-    completeness_pct: number;
-  };
-  by_justice_kind: Array<{
-    justice_kind: string;
-    justice_kind_code: string;
-    total: number;
-    has_plaintext: number;
-    has_html: number;
-    has_only_html: number;
-    has_both: number;
-    missing_both: number;
-    completeness_pct: number;
-  }>;
-}
-
-interface ScraperJob {
-  job_id: string;
-  status: 'queued' | 'running' | 'completed' | 'failed' | 'stopped';
-  justice_kind: string;
-  justice_kind_id: string;
-  doc_form: string;
-  date_from: string;
-  max_docs: number;
-  concurrency: number;
-  proxy?: string;
-  pages_processed: number;
-  downloaded: number;
-  saved_to_db: number;
-  skipped: number;
-  errors: number;
-  started_at: string;
-  completed_at?: string;
-  current_logs: string[];
-}
-
-interface CoverageMapData {
-  cells: Record<string, Record<string, number>>;
-  periods: string[];
-  justice_kinds: string[];
-  kind_labels: Record<string, string>;
-}
-
-interface BackfillJob {
-  job_id: string;
-  status: 'queued' | 'running' | 'completed' | 'failed' | 'stopped';
-  justice_kind_code: string | null;
-  total: number;
-  processed: number;
-  scraped: number;
-  errors: number;
-  error_details: string[];
-  started_at: string;
-  completed_at?: string;
-  current_logs?: string[];
-  concurrency?: number;
-  proxy?: string;
-  completeness?: {
-    summary: {
-      total_documents: number;
-      with_plaintext: number;
-      with_html: number;
-      with_only_html: number;
-      with_both: number;
-      missing_both: number;
-      completeness_pct: number;
-    };
-    by_justice_kind: Array<{
-      justice_kind: string;
-      justice_kind_code: string;
-      total: number;
-      has_plaintext: number;
-      has_html: number;
-      has_only_html: number;
-      has_both: number;
-      missing_both: number;
-      completeness_pct: number;
-    }>;
-  };
-}
-
-interface SectionState<T> {
-  data: T | null;
-  loading: boolean;
-  error: string | null;
-}
-
-interface BackendData {
-  tables: TableInfo[];
-  dbSizeMb: number;
-}
-
-interface ServiceData {
-  service: string;
-  tables: Record<string, { rows: number; source: string; sourceUrl: string; updateFrequency: string; lastUpdate: string | null; lastBatchCount?: number }>;
-  dbSizeMb: number;
-  recentImports?: any[];
-  error?: string;
-}
-
-interface ImportSample {
-  source: string;
-  source_name: string;
-  count: number;
-  last_import: string;
-    records: Array<{
-    id: string;
-    title?: string;
-    court?: string;
-    case_number?: string;
-    category?: string;
-    justice_kind?: string;
-    date?: string;
-    type?: string;
-    rada_id?: string;
-    status?: string;
-    effective_date?: string;
-    document_section_id?: string;
-    vector_id?: string;
-    document_title?: string;
-    user_email?: string;
-    user_name?: string;
-    name?: string;
-    domain?: string;
-    created_at: string;
-    updated_at?: string;
-  }>;
-}
-
-interface ImportSamplesData {
-  hours: number;
-  samples: ImportSample[];
-  summary: {
-    court_decisions: number;
-    legislation: number;
-    embeddings: number;
-    user_uploads: number;
-  };
-  timestamp: string;
-}
-
-function formatDate(dateStr: string | null): string {
-  if (!dateStr) return '—';
-  const d = new Date(dateStr);
-  if (isNaN(d.getTime())) return '—';
-  return d.toLocaleDateString('uk-UA', { day: '2-digit', month: '2-digit', year: 'numeric', hour: '2-digit', minute: '2-digit' });
-}
-
-function formatNumber(n: number | null | undefined): string {
-  if (n == null) return '0';
-  return n.toLocaleString('uk-UA');
-}
-
-function ServiceStatusBadge({ available }: { available: boolean }) {
-  return available ? (
-    <span className="inline-flex items-center gap-1.5 px-2.5 py-1 rounded-full text-xs font-medium text-green-600 bg-green-50">
-      <CheckCircle size={12} />
-      Онлайн
-    </span>
-  ) : (
-    <span className="inline-flex items-center gap-1.5 px-2.5 py-1 rounded-full text-xs font-medium text-red-600 bg-red-50">
-      <XCircle size={12} />
-      Недоступний
-    </span>
-  );
-}
-
-function SectionLoader() {
-  return (
-    <div className="bg-white rounded-xl border border-claude-border p-8 flex items-center justify-center">
-      <RefreshCw size={18} className="text-claude-subtext animate-spin mr-2" />
-      <span className="text-sm text-claude-subtext">Завантаження...</span>
-    </div>
-  );
-}
-
-function SectionError({ message, onRetry }: { message: string; onRetry: () => void }) {
-  return (
-    <div className="bg-white rounded-xl border border-claude-border p-6 text-center">
-      <XCircle size={20} className="mx-auto mb-2 text-red-400" />
-      <p className="text-sm text-red-600 mb-3">{message}</p>
-      <button onClick={onRetry} className="text-xs px-3 py-1.5 border border-claude-border rounded-lg hover:bg-gray-50 transition-colors">
-        Повторити
-      </button>
-    </div>
-  );
-}
-
-function DataTable({ tables }: { tables: TableInfo[] }) {
-  return (
-    <div className="bg-white rounded-xl border border-claude-border shadow-sm overflow-hidden">
-      <div className="overflow-x-auto">
-        <table className="w-full text-sm">
-          <thead>
-            <tr className="border-b border-claude-border bg-gray-50">
-              <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Таблиця</th>
-              <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Записів</th>
-              <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Джерело</th>
-              <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Частота оновлення</th>
-              <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Останнє оновлення</th>
-              <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Завантажено</th>
-            </tr>
-          </thead>
-          <tbody>
-            {tables.map((t) => (
-              <tr key={t.id} className="border-b border-claude-border/30 hover:bg-gray-50/50">
-                <td className="px-4 py-2.5">
-                  <div className="font-medium text-claude-text text-xs">{t.name}</div>
-                  <div className="text-[10px] text-claude-subtext font-mono">{t.id}</div>
-                </td>
-                <td className="px-4 py-2.5 text-right">
-                  <span className={`font-mono text-xs font-medium ${t.rows > 0 ? 'text-claude-text' : 'text-claude-subtext'}`}>
-                    {formatNumber(t.rows)}
-                  </span>
-                </td>
-                <td className="px-4 py-2.5">
-                  <div className="flex items-center gap-1">
-                    <span className="text-xs text-claude-text">{t.source}</span>
-                    {t.sourceUrl && (
-                      <a href={t.sourceUrl} target="_blank" rel="noopener noreferrer" className="text-blue-500 hover:text-blue-700">
-                        <ExternalLink size={10} />
-                      </a>
-                    )}
-                  </div>
-                </td>
-                <td className="px-4 py-2.5">
-                  <div className="flex items-center gap-1.5">
-                    <Clock size={11} className="text-claude-subtext flex-shrink-0" />
-                    <span className="text-xs text-claude-subtext">{t.updateFrequency}</span>
-                  </div>
-                </td>
-                <td className="px-4 py-2.5 text-xs text-claude-subtext">{formatDate(t.lastUpdate)}</td>
-                <td className="px-4 py-2.5 text-right">
-                  {t.lastBatchCount != null && t.lastBatchCount > 0 ? (
-                    <span className="inline-flex items-center gap-1 px-2 py-0.5 rounded-full text-[10px] font-medium bg-blue-50 text-blue-700 font-mono">
-                      +{formatNumber(t.lastBatchCount)}
-                    </span>
-                  ) : (
-                    <span className="text-xs text-claude-subtext">—</span>
-                  )}
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  );
-}
-
-function toTableInfoArray(tables: Record<string, any>): TableInfo[] {
-  return Object.entries(tables).map(([id, t]) => ({
-    id,
-    name: t.source?.split('—')[1]?.trim() || id,
-    rows: t.rows || 0,
-    source: t.source || '',
-    sourceUrl: t.sourceUrl || '',
-    updateFrequency: t.updateFrequency || '',
-    lastUpdate: t.lastUpdate || null,
-    lastBatchCount: t.lastBatchCount || 0,
-  }));
-}
-
-function SummaryCard({
-  icon: Icon, label, value, sub, status, loading: isLoading,
-}: {
-  icon: React.ElementType; label: string; value: string; sub: string;
-  status?: 'online' | 'offline' | 'loading'; loading?: boolean;
-}) {
-  return (
-    <div className="bg-white rounded-xl border border-claude-border p-4 shadow-sm">
-      <div className="flex items-center justify-between mb-2">
-        <div className="p-2 bg-claude-bg rounded-lg">
-          <Icon size={16} className="text-claude-text" />
-        </div>
-        {status === 'loading' || isLoading ? (
-          <RefreshCw size={10} className="text-claude-subtext animate-spin" />
-        ) : status ? (
-          <div className={`w-2 h-2 rounded-full ${status === 'online' ? 'bg-green-500' : 'bg-red-400'}`} />
-        ) : null}
-      </div>
-      <div className="text-xl font-semibold text-claude-text font-mono">
-        {isLoading ? <span className="text-claude-subtext text-base">...</span> : value}
-      </div>
-      <div className="text-xs text-claude-subtext mt-0.5">{label}</div>
-      <div className="text-[10px] text-claude-subtext/70 mt-1">{sub}</div>
-    </div>
-  );
-}
-
-function CourtDocsCategoryRow({ cat }: { cat: CourtCategory }) {
-  const [expanded, setExpanded] = useState(false);
-
-  return (
-    <>
-      <tr
-        className="border-b border-claude-border/30 hover:bg-gray-50/50 cursor-pointer"
-        onClick={() => setExpanded(!expanded)}
-      >
-        <td className="px-4 py-2.5">
-          <div className="flex items-center gap-2">
-            {expanded ? <ChevronUp size={12} className="text-claude-subtext" /> : <ChevronDown size={12} className="text-claude-subtext" />}
-            <span className="font-medium text-xs text-claude-text">{cat.name}</span>
-          </div>
-          <div className="text-[10px] text-claude-subtext font-mono ml-5">code: {cat.code}</div>
-        </td>
-        <td className="px-4 py-2.5 text-right">
-          <span className="font-mono text-xs font-medium text-claude-text">{formatNumber(cat.total)}</span>
-        </td>
-        <td className="px-4 py-2.5 text-right">
-          {cat.recent > 0 ? (
-            <span className="inline-flex items-center gap-1 px-2 py-0.5 rounded-full text-[10px] font-medium bg-green-50 text-green-700 font-mono">
-              +{formatNumber(cat.recent)}
-            </span>
-          ) : (
-            <span className="text-xs text-claude-subtext">0</span>
-          )}
-        </td>
-        <td className="px-4 py-2.5 text-xs text-claude-subtext">{formatDate(cat.last_loaded_at)}</td>
-        <td className="px-4 py-2.5 text-xs text-claude-subtext">{formatDate(cat.latest_date)}</td>
-      </tr>
-      {expanded && cat.documents.length > 0 && cat.documents.map((doc) => (
-        <tr key={doc.id} className="bg-gray-50/70 border-b border-claude-border/20">
-          <td className="px-4 py-2 pl-10" colSpan={2}>
-            <div className="text-xs text-claude-text truncate max-w-md" title={doc.title}>{doc.title}</div>
-            <div className="text-[10px] text-claude-subtext">
-              {doc.court && <span>{doc.court}</span>}
-              {doc.case_number && <span className="ml-2 font-mono">{doc.case_number}</span>}
-              {doc.dispute_category && <span className="ml-2 text-blue-600">{doc.dispute_category}</span>}
-            </div>
-          </td>
-          <td className="px-4 py-2 text-xs text-claude-subtext">{formatDate(doc.date)}</td>
-          <td className="px-4 py-2 text-xs text-claude-subtext" colSpan={2}>{formatDate(doc.loaded_at)}</td>
-        </tr>
-      ))}
-      {expanded && cat.documents.length === 0 && (
-        <tr className="bg-gray-50/70 border-b border-claude-border/20">
-          <td className="px-4 py-2 pl-10 text-xs text-claude-subtext italic" colSpan={5}>
-            Немає нових документів за обраний період
-          </td>
-        </tr>
-      )}
-    </>
-  );
-}
-
-function CourtDocsSection({
-  state,
-  onRetry,
-}: {
-  state: SectionState<CourtDocsData>;
-  onRetry: () => void;
-}) {
-  if (state.loading) return <SectionLoader />;
-  if (state.error) return <SectionError message={state.error} onRetry={onRetry} />;
-  if (!state.data) return null;
-
-  const { data } = state;
-
-  return (
-    <div>
-      {/* Summary cards */}
-      <div className="grid grid-cols-2 md:grid-cols-3 gap-3 mb-4">
-        <div className="bg-white rounded-lg border border-claude-border p-3">
-          <div className="text-lg font-semibold text-claude-text font-mono">{formatNumber(data.total_court_docs)}</div>
-          <div className="text-[10px] text-claude-subtext">Всього судових рішень</div>
-        </div>
-        <div className="bg-white rounded-lg border border-claude-border p-3">
-          <div className="text-lg font-semibold text-green-700 font-mono">+{formatNumber(data.recent_court_docs)}</div>
-          <div className="text-[10px] text-claude-subtext">За останні {data.days} днів</div>
-        </div>
-        <div className="bg-white rounded-lg border border-claude-border p-3">
-          <div className="text-lg font-semibold text-claude-text font-mono">{data.categories.length}</div>
-          <div className="text-[10px] text-claude-subtext">Видів права</div>
-        </div>
-      </div>
-
-      {/* Categories table */}
-      {data.categories.length > 0 ? (
-        <div className="bg-white rounded-xl border border-claude-border shadow-sm overflow-hidden">
-          <div className="overflow-x-auto">
-            <table className="w-full text-sm">
-              <thead>
-                <tr className="border-b border-claude-border bg-gray-50">
-                  <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Вид права</th>
-                  <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Всього</th>
-                  <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Нових</th>
-                  <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Завантажено</th>
-                  <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Дата рішення</th>
-                </tr>
-              </thead>
-              <tbody>
-                {data.categories.map((cat) => (
-                  <CourtDocsCategoryRow key={cat.code} cat={cat} />
-                ))}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      ) : (
-        <div className="bg-white rounded-xl border border-claude-border p-8 text-center">
-          <FileText size={24} className="mx-auto mb-2 text-claude-subtext" />
-          <p className="text-sm text-claude-subtext">Судові документи ще не завантажені</p>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function completenessColor(pct: number): string {
-  if (pct >= 100) return 'text-green-700';
-  if (pct >= 80) return 'text-yellow-600';
-  return 'text-red-600';
-}
-
-function completenessBg(pct: number): string {
-  if (pct >= 100) return 'bg-green-50';
-  if (pct >= 80) return 'bg-yellow-50';
-  return 'bg-red-50';
-}
-
-function BackfillProgress({ job, onStop, onRefresh, onDelete }: { job: BackfillJob; onStop: () => void; onRefresh: () => void; onDelete: () => void }) {
-  const isActive = job.status === 'running' || job.status === 'queued';
-  const progressPct = job.total > 0 ? Math.round((job.processed / job.total) * 100) : 0;
-
-  return (
-    <div className="bg-blue-50 border border-blue-200 rounded-lg p-4 mb-4">
-      <div className="flex items-center justify-between mb-2">
-        <div className="flex items-center gap-2">
-          {isActive ? (
-            <RefreshCw size={14} className="text-blue-600 animate-spin" />
-          ) : job.status === 'completed' ? (
-            <CheckCircle size={14} className="text-green-600" />
-          ) : job.status === 'stopped' ? (
-            <Square size={14} className="text-yellow-600" />
-          ) : (
-            <XCircle size={14} className="text-red-600" />
-          )}
-          <span className="text-sm font-medium text-blue-900">
-            {isActive ? 'Докачування...' : job.status === 'completed' ? 'Завершено' : job.status === 'stopped' ? 'Зупинено' : 'Помилка'}
-          </span>
-        </div>
-        <div className="flex items-center gap-2">
-          {isActive && (
-            <button
-              onClick={onStop}
-              className="flex items-center gap-1 px-2 py-1 text-xs text-red-700 bg-red-100 border border-red-200 rounded hover:bg-red-200 transition-colors"
-            >
-              <Square size={10} />
-              Зупинити
-            </button>
-          )}
-          {!isActive && (
-            <>
-              <button
-                onClick={onDelete}
-                className="flex items-center gap-1 px-2 py-1 text-xs text-gray-500 bg-gray-100 border border-gray-200 rounded hover:bg-gray-200 transition-colors"
-                title="Видалити задачу"
-              >
-                <X size={10} /> Видалити
-              </button>
-              <button
-                onClick={onRefresh}
-                className="text-xs text-blue-600 hover:underline"
-              >
-                Приховати
-              </button>
-            </>
-          )}
-        </div>
-      </div>
-
-      {/* Progress bar */}
-      <div className="w-full bg-blue-200 rounded-full h-2 mb-2">
-        <div
-          className={`h-2 rounded-full transition-all duration-500 ${job.status === 'completed' ? 'bg-green-500' : job.status === 'failed' ? 'bg-red-500' : 'bg-blue-600'}`}
-          style={{ width: `${progressPct}%` }}
-        />
-      </div>
-
-      <div className="flex items-center gap-4 text-xs text-blue-800">
-        <span>{job.processed}/{job.total} оброблено</span>
-        <span className="text-green-700">{job.scraped} докачано</span>
-        {job.errors > 0 && <span className="text-red-600">{job.errors} помилок</span>}
-        <span className="text-blue-600">{progressPct}%</span>
-        {(job.concurrency ?? 1) > 1 && <span className="text-purple-700">{job.concurrency} потоків</span>}
-        {job.proxy && <span className="text-orange-700">проксі</span>}
-      </div>
-
-      {isActive && job.current_logs && job.current_logs.length > 0 && (
-        <div className="mt-3 p-2 bg-slate-900 rounded text-[10px] font-mono text-green-400">
-          {job.current_logs.map((log, i) => (
-            <div key={i} className="truncate">{log}</div>
-          ))}
-        </div>
-      )}
-
-      {job.error_details.length > 0 && !isActive && (
-        <details className="mt-2">
-          <summary className="text-xs text-red-600 cursor-pointer">Деталі помилок ({job.error_details.length})</summary>
-          <div className="mt-1 max-h-32 overflow-y-auto text-[10px] text-red-700 font-mono bg-red-50 p-2 rounded">
-            {job.error_details.map((e, i) => <div key={i}>{e}</div>)}
-          </div>
-        </details>
-      )}
-    </div>
-  );
-}
-
-const JUSTICE_KINDS = [
-  { label: 'Кримінальне', value: 'Кримінальне', id: '5' },
-  { label: 'Цивільне', value: 'Цивільне', id: '1' },
-  { label: 'Адміністративне', value: 'Адміністративне', id: '2' },
-  { label: 'Господарське', value: 'Господарське', id: '3' },
-];
-
-const DOC_FORMS = [
-  { label: 'Усі форми', value: '__all__' },
-  { label: 'Рішення', value: 'Рішення' },
-  { label: 'Ухвала', value: 'Ухвала' },
-  { label: 'Вирок', value: 'Вирок' },
-  { label: 'Постанова', value: 'Постанова' },
-];
-
-const MONTHS_SHORT = ['С','Л','Б','К','Т','Ч','Л','С','В','Ж','Л','Г'];
-
-function ScraperJobCard({ job, onStop, onDelete }: { job: ScraperJob; onStop: (id: string) => void; onDelete: (id: string) => void }) {
-  const isActive = job.status === 'running' || job.status === 'queued';
-  const [logsOpen, setLogsOpen] = useState(false);
-  return (
-    <div className={`rounded-lg border p-3 ${isActive ? 'border-blue-200 bg-blue-50' : job.status === 'completed' ? 'border-green-100 bg-green-50' : job.status === 'stopped' ? 'border-yellow-100 bg-yellow-50' : 'border-red-100 bg-red-50'}`}>
-      <div className="flex items-center justify-between mb-1.5">
-        <div className="flex items-center gap-2 flex-wrap">
-          {isActive ? (
-            <RefreshCw size={13} className="text-blue-600 animate-spin shrink-0" />
-          ) : job.status === 'completed' ? (
-            <CheckCircle size={13} className="text-green-600 shrink-0" />
-          ) : job.status === 'stopped' ? (
-            <Square size={13} className="text-yellow-600 shrink-0" />
-          ) : (
-            <XCircle size={13} className="text-red-600 shrink-0" />
-          )}
-          <span className="text-xs font-medium text-gray-800">{job.justice_kind}</span>
-          <span className="text-[10px] text-gray-500">від {job.date_from}</span>
-          {job.proxy ? (
-            <span className="inline-flex px-1.5 py-0.5 rounded-full text-[9px] bg-orange-100 text-orange-700 font-medium">Proxy</span>
-          ) : (
-            <span className="inline-flex px-1.5 py-0.5 rounded-full text-[9px] bg-gray-100 text-gray-500">Direct</span>
-          )}
-          <span className="text-[10px] text-gray-400">{job.concurrency}×</span>
-        </div>
-        <div className="flex items-center gap-2 shrink-0">
-          {isActive && (
-            <button
-              onClick={() => onStop(job.job_id)}
-              className="flex items-center gap-1 px-2 py-0.5 text-[10px] text-red-700 bg-red-100 border border-red-200 rounded hover:bg-red-200 transition-colors"
-            >
-              <Square size={9} /> Стоп
-            </button>
-          )}
-          {!isActive && (
-            <button
-              onClick={() => onDelete(job.job_id)}
-              className="flex items-center gap-1 px-2 py-0.5 text-[10px] text-gray-500 bg-gray-100 border border-gray-200 rounded hover:bg-gray-200 transition-colors"
-              title="Видалити задачу"
-            >
-              <X size={9} /> Видалити
-            </button>
-          )}
-          {job.current_logs && job.current_logs.length > 0 && (
-            <button onClick={() => setLogsOpen(v => !v)} className="text-[10px] text-blue-600 hover:underline">
-              {logsOpen ? 'Сховати' : 'Логи'}
-            </button>
-          )}
-        </div>
-      </div>
-      <div className="flex items-center gap-3 text-[10px] font-mono">
-        <span className="text-gray-500">Стор: <span className="text-gray-800">{job.pages_processed}</span></span>
-        <span className="text-gray-500">↓ <span className="text-blue-700">{job.downloaded}</span></span>
-        <span className="text-gray-500">БД: <span className="text-green-700">{job.saved_to_db}</span></span>
-        <span className="text-gray-500">Пропуск: <span className="text-yellow-700">{job.skipped}</span></span>
-        {job.errors > 0 && <span className="text-red-600">Помилок: {job.errors}</span>}
-      </div>
-      {logsOpen && job.current_logs && job.current_logs.length > 0 && (
-        <div className="mt-2 p-2 bg-slate-900 rounded text-[10px] font-mono text-green-400 max-h-28 overflow-y-auto">
-          {job.current_logs.map((log, i) => <div key={i} className="truncate">{log}</div>)}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function CourtDataMapSection() {
-  const [jobs, setJobs] = useState<ScraperJob[]>([]);
-  const [mapData, setMapData] = useState<CoverageMapData | null>(null);
-  const [mapYears, setMapYears] = useState(3);
-  const [mapLoading, setMapLoading] = useState(false);
-  const [showAddBatch, setShowAddBatch] = useState(false);
-  const [starting, setStarting] = useState(false);
-  const [config, setConfig] = useState({
-    justice_kind: 'Кримінальне',
-    justice_kind_id: '5',
-    doc_form: '__all__',
-    date_from: '01.01.2020',
-    max_docs: 10000,
-    concurrency: 5,
-    proxy: 'none',
-  });
-  const pollRef = React.useRef<ReturnType<typeof setInterval> | null>(null);
-
-  const loadJobs = useCallback(async () => {
-    try {
-      const res = await api.admin.getAllScraperJobs();
-      setJobs(res.data.jobs || []);
-    } catch { /* ignore */ }
-  }, []);
-
-  const loadMap = useCallback(async (years: number) => {
-    setMapLoading(true);
-    try {
-      const res = await api.admin.getRegistryCoverageMap(years);
-      setMapData(res.data);
-    } catch { /* ignore */ } finally {
-      setMapLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    loadMap(mapYears);
-    loadJobs();
-    return () => { if (pollRef.current) clearInterval(pollRef.current); };
-  }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-  useEffect(() => {
-    if (pollRef.current) clearInterval(pollRef.current);
-    const hasActive = jobs.some(j => j.status === 'running' || j.status === 'queued');
-    if (hasActive) {
-      pollRef.current = setInterval(() => loadJobs(), 2500);
-    }
-    return () => { if (pollRef.current) clearInterval(pollRef.current); };
-  }, [jobs, loadJobs]);
-
-  const handleCellClick = (jk: string, period: string) => {
-    const [year, month] = period.split('-');
-    const date_from = `01.${month}.${year}`;
-    const found = JUSTICE_KINDS.find(k => k.id === jk);
-    if (found) {
-      setConfig(c => ({ ...c, justice_kind: found.value, justice_kind_id: found.id, date_from }));
-    } else {
-      setConfig(c => ({ ...c, date_from }));
-    }
-    setShowAddBatch(true);
-  };
-
-  const handleStart = async () => {
-    setStarting(true);
-    try {
-      await api.admin.startCourtScraper(config);
-      setShowAddBatch(false);
-      await loadJobs();
-    } catch (e: any) {
-      alert(e?.response?.data?.error || 'Помилка запуску');
-    } finally {
-      setStarting(false);
-    }
-  };
-
-  const handleStop = async (jobId: string) => {
-    try { await api.admin.stopCourtScraper(jobId); await loadJobs(); } catch { /* ignore */ }
-  };
-
-  const handleDelete = async (jobId: string) => {
-    try { await api.admin.deleteCourtScraperJob(jobId); await loadJobs(); } catch { /* ignore */ }
-  };
-
-  // Group periods by year for header rendering
-  const periodsByYear = React.useMemo(() => {
-    if (!mapData) return {} as Record<string, string[]>;
-    const groups: Record<string, string[]> = {};
-    for (const p of mapData.periods) {
-      const year = p.split('-')[0];
-      if (!groups[year]) groups[year] = [];
-      groups[year].push(p);
-    }
-    return groups;
-  }, [mapData]);
-
-  const sortedYears = Object.keys(periodsByYear).sort();
-  const activeCount = jobs.filter(j => j.status === 'running' || j.status === 'queued').length;
-
-  // Build set of cell keys that have active (running/queued) jobs → used for blink animation
-  const activeCellKeys = React.useMemo(() => {
-    const keys = new Set<string>();
-    for (const job of jobs) {
-      if (job.status !== 'running' && job.status !== 'queued') continue;
-      // date_from format: dd.mm.yyyy → convert to YYYY-MM
-      const parts = job.date_from.split('.');
-      if (parts.length === 3) {
-        const period = `${parts[2]}-${parts[1]}`;
-        keys.add(`${job.justice_kind_id}-${period}`);
-      }
-    }
-    return keys;
-  }, [jobs]);
-
-  return (
-    <div className="space-y-4">
-      {/* Coverage Map Card */}
-      <div className="bg-white rounded-xl border border-claude-border p-5">
-        <div className="flex items-center justify-between mb-3">
-          <div className="flex items-center gap-2">
-            <Grid3X3 size={15} className="text-claude-subtext" />
-            <span className="text-sm font-medium text-claude-text">Карта покриття реєстру</span>
-            <select
-              value={mapYears}
-              onChange={e => { const y = Number(e.target.value); setMapYears(y); loadMap(y); }}
-              className="text-xs border border-claude-border rounded px-1.5 py-0.5 bg-white text-claude-text"
-            >
-              <option value={2}>2 роки</option>
-              <option value={3}>3 роки</option>
-              <option value={5}>5 років</option>
-            </select>
-            <button onClick={() => loadMap(mapYears)} title="Оновити" className="text-claude-subtext hover:text-claude-text transition-colors">
-              <RefreshCw size={13} className={mapLoading ? 'animate-spin' : ''} />
-            </button>
-          </div>
-          <button
-            onClick={() => setShowAddBatch(v => !v)}
-            className="flex items-center gap-1.5 px-3 py-1.5 bg-blue-600 text-white text-xs rounded-lg hover:bg-blue-700 transition-colors"
-          >
-            <Plus size={13} />
-            Новий батч
-            {activeCount > 0 && (
-              <span className="ml-1 bg-blue-400 text-white rounded-full px-1.5 py-0.5 text-[9px] font-bold">{activeCount}</span>
-            )}
-          </button>
-        </div>
-
-        {mapLoading && !mapData ? (
-          <div className="h-24 flex items-center justify-center text-gray-400">
-            <RefreshCw size={20} className="animate-spin" />
-          </div>
-        ) : mapData && mapData.justice_kinds.length > 0 ? (
-          <div className="overflow-x-auto">
-            <div className="min-w-max">
-              {/* Year header row */}
-              <div className="flex mb-0.5">
-                <div className="w-32 shrink-0" />
-                {sortedYears.map(year => (
-                  <div
-                    key={year}
-                    style={{ width: `${periodsByYear[year].length * 18}px` }}
-                    className="text-center text-[9px] text-gray-400 font-medium border-l border-gray-200 px-1 overflow-hidden"
-                  >
-                    {year}
-                  </div>
-                ))}
-              </div>
-
-              {/* Month labels */}
-              <div className="flex items-center mb-1.5">
-                <div className="w-32 shrink-0" />
-                {mapData.periods.map(period => {
-                  const month = parseInt(period.split('-')[1]) - 1;
-                  return (
-                    <div key={period} className="w-[16px] mr-[2px] text-[8px] text-center text-gray-300 leading-none">
-                      {MONTHS_SHORT[month]}
-                    </div>
-                  );
-                })}
-              </div>
-
-              {/* Data rows: one per justice kind */}
-              {mapData.justice_kinds.map(jk => (
-                <div key={jk} className="flex items-center mb-[3px]">
-                  <div className="w-32 text-[11px] text-right pr-3 text-gray-600 shrink-0 truncate" title={mapData.kind_labels[jk] || jk}>
-                    {mapData.kind_labels[jk] || jk}
-                  </div>
-                  {mapData.periods.map(period => {
-                    const count = mapData.cells[jk]?.[period] ?? 0;
-                    const isActive = activeCellKeys.has(`${jk}-${period}`);
-                    return (
-                      <div
-                        key={period}
-                        onClick={() => handleCellClick(jk, period)}
-                        title={`${mapData.kind_labels[jk] || jk} · ${period}: ${count.toLocaleString('uk')} документів\nКлік — запустити батч`}
-                        className={`w-[16px] h-[16px] mr-[2px] rounded-[2px] cursor-pointer hover:ring-2 hover:ring-blue-400 hover:ring-offset-0 transition-all ${
-                          isActive ? 'animate-pulse ring-2 ring-amber-400 ring-offset-0' : ''
-                        } ${
-                          count === 0 ? 'bg-gray-100 hover:bg-gray-200' :
-                          count < 50 ? 'bg-sky-100' :
-                          count < 200 ? 'bg-sky-300' :
-                          count < 1000 ? 'bg-sky-500' :
-                          count < 5000 ? 'bg-sky-700' :
-                          'bg-sky-900'
-                        }`}
-                      />
-                    );
-                  })}
-                </div>
-              ))}
-
-              {/* Legend */}
-              <div className="flex items-center gap-1.5 mt-3 ml-32">
-                <span className="text-[9px] text-gray-400">0</span>
-                {['bg-gray-100', 'bg-sky-100', 'bg-sky-300', 'bg-sky-500', 'bg-sky-700', 'bg-sky-900'].map(cls => (
-                  <div key={cls} className={`w-3 h-3 rounded-[2px] ${cls} border border-gray-100`} />
-                ))}
-                <span className="text-[9px] text-gray-400">5000+</span>
-                <span className="text-[9px] text-gray-300 ml-3">Клік на клітинку → запустити батч</span>
-              </div>
-            </div>
-          </div>
-        ) : !mapLoading ? (
-          <div className="h-16 flex items-center justify-center text-xs text-gray-400">
-            Даних для карти немає — запустіть перший батч скрапінгу
-          </div>
-        ) : null}
-      </div>
-
-      {/* Add Batch Form */}
-      {showAddBatch && (
-        <div className="bg-blue-50 border border-blue-200 rounded-xl p-4">
-          <div className="flex items-center justify-between mb-3">
-            <span className="text-sm font-medium text-blue-900">Новий батч скрапінгу</span>
-            <button onClick={() => setShowAddBatch(false)} className="text-blue-400 hover:text-blue-600 transition-colors">
-              <X size={16} />
-            </button>
-          </div>
-          <div className="grid grid-cols-2 md:grid-cols-3 gap-3 mb-3">
-            <div>
-              <label className="block text-xs text-blue-700 mb-1">Вид судочинства</label>
-              <select
-                className="w-full text-sm border border-blue-200 rounded px-2 py-1.5 bg-white"
-                value={config.justice_kind}
-                onChange={e => {
-                  const found = JUSTICE_KINDS.find(k => k.value === e.target.value);
-                  setConfig(c => ({ ...c, justice_kind: e.target.value, justice_kind_id: found?.id || '1' }));
-                }}
-              >
-                {JUSTICE_KINDS.map(k => <option key={k.value} value={k.value}>{k.label}</option>)}
-              </select>
-            </div>
-            <div>
-              <label className="block text-xs text-blue-700 mb-1">Форма документа</label>
-              <select
-                className="w-full text-sm border border-blue-200 rounded px-2 py-1.5 bg-white"
-                value={config.doc_form}
-                onChange={e => setConfig(c => ({ ...c, doc_form: e.target.value }))}
-              >
-                {DOC_FORMS.map(f => <option key={f.value} value={f.value}>{f.label}</option>)}
-              </select>
-            </div>
-            <div>
-              <label className="block text-xs text-blue-700 mb-1">Дата від</label>
-              <input
-                type="text"
-                className="w-full text-sm border border-blue-200 rounded px-2 py-1.5 bg-white"
-                placeholder="дд.мм.рррр"
-                value={config.date_from}
-                onChange={e => setConfig(c => ({ ...c, date_from: e.target.value }))}
-              />
-            </div>
-            <div>
-              <label className="block text-xs text-blue-700 mb-1">Макс. документів</label>
-              <input
-                type="number"
-                className="w-full text-sm border border-blue-200 rounded px-2 py-1.5 bg-white"
-                min={100} max={50000} step={100}
-                value={config.max_docs}
-                onChange={e => setConfig(c => ({ ...c, max_docs: parseInt(e.target.value) || 10000 }))}
-              />
-            </div>
-            <div>
-              <label className="block text-xs text-blue-700 mb-1">Потоків: {config.concurrency}</label>
-              <input
-                type="range" min={1} max={8} step={1}
-                value={config.concurrency}
-                onChange={e => setConfig(c => ({ ...c, concurrency: parseInt(e.target.value) }))}
-                className="w-full mt-1.5"
-              />
-            </div>
-            <div>
-              <label className="block text-xs text-blue-700 mb-1">З'єднання</label>
-              <select
-                className="w-full text-sm border border-blue-200 rounded px-2 py-1.5 bg-white"
-                value={config.proxy}
-                onChange={e => setConfig(c => ({ ...c, proxy: e.target.value }))}
-              >
-                <option value="none">Без проксі (Direct)</option>
-                <option value="mail">Проксі: mail.legal.org.ua</option>
-                <option value="localdev">Проксі: localdev.legal.org.ua</option>
-              </select>
-            </div>
-          </div>
-          <button
-            onClick={handleStart}
-            disabled={starting}
-            className="flex items-center gap-2 px-4 py-2 bg-blue-600 text-white text-sm rounded-lg hover:bg-blue-700 disabled:opacity-50 transition-colors"
-          >
-            <Play size={14} />
-            {starting ? 'Запуск...' : 'Запустити батч'}
-          </button>
-        </div>
-      )}
-
-      {/* Active & Recent Jobs */}
-      {jobs.length > 0 && (
-        <div className="space-y-2">
-          <div className="flex items-center justify-between px-1">
-            <span className="text-xs font-medium text-claude-subtext">
-              Батчі ({activeCount} активних з {jobs.length})
-            </span>
-            <button onClick={loadJobs} className="text-claude-subtext hover:text-claude-text transition-colors" title="Оновити">
-              <RefreshCw size={11} />
-            </button>
-          </div>
-          {jobs.slice(0, 10).map(job => (
-            <ScraperJobCard key={job.job_id} job={job} onStop={handleStop} onDelete={handleDelete} />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function DocumentCompletenessSection() {
-  const [result, setResult] = useState<CompletenessResult | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [limitReached, setLimitReached] = useState(false);
-
-  // Backfill state
-  const [backfillJob, setBackfillJob] = useState<BackfillJob | null>(null);
-  const [backfillStarting, setBackfillStarting] = useState(false);
-  const [backfillConfig, setBackfillConfig] = useState({
-    justice_kind_code: 'all',
-    limit: 200,
-    concurrency: 1,
-    proxy: 'none',
-  });
-  const pollRef = React.useRef<ReturnType<typeof setInterval> | null>(null);
-
-  const stopPolling = useCallback(() => {
-    if (pollRef.current) {
-      clearInterval(pollRef.current);
-      pollRef.current = null;
-    }
-  }, []);
-
-  const pollBackfillStatus = useCallback((jobId: string) => {
-    stopPolling();
-    pollRef.current = setInterval(async () => {
-      try {
-        const res = await api.admin.getBackfillStatus(jobId);
-        const job = res.data;
-        setBackfillJob(job);
-        if (job.status !== 'running' && job.status !== 'queued') {
-          stopPolling();
-        }
-      } catch {
-        stopPolling();
-      }
-    }, 2000);
-  }, [stopPolling]);
-
-  // Check for active backfill on mount
-  useEffect(() => {
-    (async () => {
-      try {
-        const res = await api.admin.getBackfillStatus();
-        if (res.data.active && res.data.job) {
-          setBackfillJob(res.data.job);
-          pollBackfillStatus(res.data.job.job_id);
-        } else if (res.data.job && res.data.job.status !== 'running' && res.data.job.status !== 'queued') {
-          // Show last completed job briefly
-          setBackfillJob(res.data.job);
-        }
-      } catch { /* no active job */ }
-    })();
-    return stopPolling;
-  }, [pollBackfillStatus, stopPolling]);
-
-  const runCheck = async () => {
-    setLoading(true);
-    setError(null);
-    try {
-      const res = await api.admin.runDocumentCompletenessCheck();
-      setResult(res.data);
-      if (res.data.runs_today >= res.data.max_runs_per_day) {
-        setLimitReached(true);
-      }
-    } catch (err: any) {
-      if (err.response?.status === 429) {
-        setLimitReached(true);
-        setError(err.response?.data?.error || 'Ліміт вичерпано');
-      } else {
-        setError(err.response?.data?.error || err.message);
-      }
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const startBackfill = async (justiceKindCode?: string) => {
-    setBackfillStarting(true);
-    try {
-      const res = await api.admin.startBackfillFulltext({
-        justice_kind_code: justiceKindCode || backfillConfig.justice_kind_code,
-        limit: backfillConfig.limit,
-        concurrency: backfillConfig.concurrency,
-        proxy: backfillConfig.proxy,
-      });
-      const job = { ...res.data, processed: 0, scraped: 0, errors: 0, error_details: [], started_at: new Date().toISOString() } as BackfillJob;
-      setBackfillJob(job);
-      pollBackfillStatus(res.data.job_id);
-    } catch (err: any) {
-      if (err.response?.status === 409) {
-        // Already running — fetch status
-        try {
-          const statusRes = await api.admin.getBackfillStatus();
-          if (statusRes.data.job) {
-            setBackfillJob(statusRes.data.job);
-            pollBackfillStatus(statusRes.data.job.job_id);
-          }
-        } catch { /* ignore */ }
-      } else {
-        setError(err.response?.data?.error || err.message);
-      }
-    } finally {
-      setBackfillStarting(false);
-    }
-  };
-
-  const stopBackfill = async () => {
-    if (!backfillJob) return;
-    try {
-      await api.admin.stopBackfill(backfillJob.job_id);
-    } catch { /* ignore */ }
-  };
-
-  const deleteBackfill = async () => {
-    if (!backfillJob) return;
-    try {
-      await api.admin.deleteBackfillJob(backfillJob.job_id);
-      setBackfillJob(null);
-      stopPolling();
-    } catch { /* ignore */ }
-  };
-
-  const isBackfillActive = backfillJob && (backfillJob.status === 'running' || backfillJob.status === 'queued');
-
-  const displayData = backfillJob?.completeness ?? result;
-
-  return (
-    <div>
-      {/* Action buttons */}
-      <div className="flex items-center gap-3 mb-4">
-        <button
-          onClick={runCheck}
-          disabled={loading || limitReached}
-          className="flex items-center gap-2 px-4 py-2 bg-white border border-claude-border rounded-lg text-sm text-claude-text hover:bg-claude-bg transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-        >
-          {loading ? (
-            <RefreshCw size={14} className="animate-spin" />
-          ) : (
-            <Play size={14} />
-          )}
-          {limitReached ? 'Ліміт вичерпано' : 'Запустити перевірку'}
-        </button>
-
-        {result && result.summary.missing_both > 0 && !isBackfillActive && (
-          <button
-            onClick={() => startBackfill()}
-            disabled={backfillStarting}
-            className="flex items-center gap-2 px-4 py-2 bg-blue-600 text-white rounded-lg text-sm hover:bg-blue-700 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-          >
-            {backfillStarting ? (
-              <RefreshCw size={14} className="animate-spin" />
-            ) : (
-              <Download size={14} />
-            )}
-            Докачати всі ({formatNumber(result.summary.missing_both)})
-          </button>
-        )}
-
-        {result && (
-          <span className="text-xs text-claude-subtext">
-            {result.runs_today}/{result.max_runs_per_day} перевірок сьогодні · {formatDate(result.checked_at)}
-          </span>
-        )}
-      </div>
-
-      {/* Backfill Configuration */}
-      {!isBackfillActive && (
-        <div className="bg-gray-50 border border-claude-border rounded-lg p-4 mb-4">
-          <div className="text-sm font-medium text-claude-text mb-3">Конфігурація докачування</div>
-          <div className="grid grid-cols-2 md:grid-cols-4 gap-4">
-            <div>
-              <label className="block text-xs text-claude-subtext mb-1">Вид права</label>
-              <select
-                value={backfillConfig.justice_kind_code}
-                onChange={(e) => setBackfillConfig(c => ({ ...c, justice_kind_code: e.target.value }))}
-                className="w-full text-xs border border-claude-border rounded px-2 py-1.5 bg-white text-claude-text"
-              >
-                <option value="all">Всі види</option>
-                {result?.by_justice_kind.map(jk => (
-                  <option key={jk.justice_kind_code} value={jk.justice_kind_code}>
-                    {jk.justice_kind} ({formatNumber(jk.missing_both)})
-                  </option>
-                ))}
-              </select>
-            </div>
-            <div>
-              <label className="block text-xs text-claude-subtext mb-1">Лімит документів</label>
-              <input
-                type="number"
-                min={1}
-                max={1000}
-                value={backfillConfig.limit}
-                onChange={(e) => setBackfillConfig(c => ({ ...c, limit: Math.min(1000, Math.max(1, parseInt(e.target.value) || 200)) }))}
-                className="w-full text-xs border border-claude-border rounded px-2 py-1.5 bg-white text-claude-text"
-              />
-            </div>
-            <div>
-              <label className="block text-xs text-claude-subtext mb-1">Потоків: {backfillConfig.concurrency}</label>
-              <input
-                type="range"
-                min={1}
-                max={10}
-                value={backfillConfig.concurrency}
-                onChange={(e) => setBackfillConfig(c => ({ ...c, concurrency: parseInt(e.target.value) }))}
-                className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer"
-              />
-            </div>
-            <div>
-              <label className="block text-xs text-claude-subtext mb-1">Проксі</label>
-              <select
-                value={backfillConfig.proxy}
-                onChange={(e) => setBackfillConfig(c => ({ ...c, proxy: e.target.value }))}
-                className="w-full text-xs border border-claude-border rounded px-2 py-1.5 bg-white text-claude-text"
-              >
-                <option value="none">Без проксі</option>
-                <option value="mail">Mail Server (порт 8888)</option>
-                <option value="localdev">LocalDev (порт 8888)</option>
-              </select>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Backfill progress */}
-      {backfillJob && (
-        <BackfillProgress
-          job={backfillJob}
-          onStop={stopBackfill}
-          onRefresh={() => setBackfillJob(null)}
-          onDelete={deleteBackfill}
-        />
-      )}
-
-      {error && !result && (
-        <div className="bg-red-50 border border-red-200 rounded-lg p-3 mb-4 text-sm text-red-700">{error}</div>
-      )}
-
-      {(result || (backfillJob?.completeness)) && (
-        <>
-          {/* Summary cards */}
-          <div className="grid grid-cols-2 md:grid-cols-5 gap-3 mb-4">
-            <div className="bg-white rounded-lg border border-claude-border p-3">
-              <div className="text-lg font-semibold text-claude-text font-mono">
-                {formatNumber(backfillJob?.completeness?.summary.total_documents ?? result?.summary.total_documents)}
-              </div>
-              <div className="text-[10px] text-claude-subtext">Всього документів</div>
-            </div>
-            <div className={`rounded-lg border border-claude-border p-3 ${completenessBg(backfillJob?.completeness?.summary.completeness_pct ?? result?.summary.completeness_pct ?? 0)}`}>
-              <div className={`text-lg font-semibold font-mono ${completenessColor(backfillJob?.completeness?.summary.completeness_pct ?? result?.summary.completeness_pct ?? 0)}`}>
-                {backfillJob?.completeness?.summary.completeness_pct ?? result?.summary.completeness_pct ?? 0}%
-              </div>
-              <div className="text-[10px] text-claude-subtext">Повнота (обидва поля)</div>
-            </div>
-            <div className="bg-white rounded-lg border border-claude-border p-3">
-              <div className="text-lg font-semibold text-green-700 font-mono">
-                {formatNumber(backfillJob?.completeness?.summary.with_both ?? result?.summary.with_both ?? 0)}
-              </div>
-              <div className="text-[10px] text-claude-subtext">З обома полями</div>
-            </div>
-            <div className="bg-white rounded-lg border border-claude-border p-3">
-              <div className="text-lg font-semibold text-orange-700 font-mono">
-                {formatNumber(backfillJob?.completeness?.summary.with_only_html ?? result?.summary.with_only_html ?? 0)}
-              </div>
-              <div className="text-[10px] text-claude-subtext">Тільки HTML</div>
-            </div>
-            <div className={`rounded-lg border border-claude-border p-3 ${(backfillJob?.completeness?.summary.missing_both ?? result?.summary.missing_both ?? 0) > 0 ? 'bg-red-50' : 'bg-white'}`}>
-              <div className={`text-lg font-semibold font-mono ${(backfillJob?.completeness?.summary.missing_both ?? result?.summary.missing_both ?? 0) > 0 ? 'text-red-600' : 'text-claude-text'}`}>
-                {formatNumber(backfillJob?.completeness?.summary.missing_both ?? result?.summary.missing_both ?? 0)}
-              </div>
-              <div className="text-[10px] text-claude-subtext">Без обох полів</div>
-            </div>
-          </div>
-
-          {/* Breakdown table */}
-          <div className="bg-white rounded-xl border border-claude-border shadow-sm overflow-hidden">
-            <div className="overflow-x-auto">
-              <table className="w-full text-sm">
-                <thead>
-                  <tr className="border-b border-claude-border bg-gray-50">
-                    <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Вид права</th>
-                    <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Всього</th>
-                    <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Plaintext</th>
-                    <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Тільки HTML</th>
-                    <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Обидва</th>
-                    <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Відсутні</th>
-                    <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Повнота %</th>
-                    <th className="text-center px-4 py-2.5 font-medium text-claude-subtext text-xs">Дії</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {displayData?.by_justice_kind.map((row) => (
-                    <tr key={row.justice_kind_code} className="border-b border-claude-border/30 hover:bg-gray-50/50">
-                      <td className="px-4 py-2.5">
-                        <div className="font-medium text-xs text-claude-text">{row.justice_kind}</div>
-                        <div className="text-[10px] text-claude-subtext font-mono">{row.justice_kind_code}</div>
-                      </td>
-                      <td className="px-4 py-2.5 text-right font-mono text-xs">{formatNumber(row.total)}</td>
-                      <td className="px-4 py-2.5 text-right font-mono text-xs">{formatNumber(row.has_plaintext)}</td>
-                      <td className="px-4 py-2.5 text-right font-mono text-xs text-orange-700">{formatNumber(row.has_only_html ?? 0)}</td>
-                      <td className="px-4 py-2.5 text-right font-mono text-xs text-green-700">{formatNumber(row.has_both)}</td>
-                      <td className="px-4 py-2.5 text-right font-mono text-xs">
-                        <span className={row.missing_both > 0 ? 'text-red-600' : ''}>{formatNumber(row.missing_both)}</span>
-                      </td>
-                      <td className="px-4 py-2.5 text-right">
-                        <span className={`inline-flex px-2 py-0.5 rounded-full text-[10px] font-medium font-mono ${completenessBg(row.completeness_pct)} ${completenessColor(row.completeness_pct)}`}>
-                          {row.completeness_pct}%
-                        </span>
-                      </td>
-                      <td className="px-4 py-2.5 text-center">
-                        {row.missing_both > 0 && !isBackfillActive && (
-                          <button
-                            onClick={() => startBackfill(row.justice_kind_code)}
-                            disabled={backfillStarting}
-                            className="inline-flex items-center gap-1 px-2 py-1 text-[10px] font-medium text-blue-700 bg-blue-50 border border-blue-200 rounded hover:bg-blue-100 transition-colors disabled:opacity-50"
-                            title={`Докачати ${row.missing_both} документів`}
-                          >
-                            <Download size={10} />
-                            Докачати
-                          </button>
-                        )}
-                      </td>
-                    </tr>
-                  ))}
-                  {/* Summary row */}
-                  <tr className="border-t-2 border-claude-border bg-gray-50 font-semibold">
-                    <td className="px-4 py-2.5 text-xs text-claude-text">Всього</td>
-                    <td className="px-4 py-2.5 text-right font-mono text-xs">{formatNumber(displayData?.summary.total_documents ?? 0)}</td>
-                    <td className="px-4 py-2.5 text-right font-mono text-xs">{formatNumber(displayData?.summary.with_plaintext ?? 0)}</td>
-                    <td className="px-4 py-2.5 text-right font-mono text-xs text-orange-700">{formatNumber(displayData?.summary.with_only_html ?? 0)}</td>
-                    <td className="px-4 py-2.5 text-right font-mono text-xs text-green-700">{formatNumber(displayData?.summary.with_both ?? 0)}</td>
-                    <td className="px-4 py-2.5 text-right font-mono text-xs">
-                      <span className={(displayData?.summary.missing_both ?? 0) > 0 ? 'text-red-600' : ''}>{formatNumber(displayData?.summary.missing_both ?? 0)}</span>
-                    </td>
-                    <td className="px-4 py-2.5 text-right">
-                      <span className={`inline-flex px-2 py-0.5 rounded-full text-[10px] font-medium font-mono ${completenessBg(displayData?.summary.completeness_pct ?? 0)} ${completenessColor(displayData?.summary.completeness_pct ?? 0)}`}>
-                        {displayData?.summary.completeness_pct ?? 0}%
-                      </span>
-                    </td>
-                    <td></td>
-                  </tr>
-                </tbody>
-              </table>
-            </div>
-          </div>
-        </>
-      )}
-    </div>
-  );
-}
-
-
-interface DocumentModalProps {
-  documentId: string | null;
-  isOpen: boolean;
-  onClose: () => void;
-}
-
-function DocumentModal({ documentId, isOpen, onClose }: DocumentModalProps) {
-  const [document, setDocument] = useState<any>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    if (documentId && isOpen) {
-      fetchDocument();
-    }
-  }, [documentId, isOpen]);
-
-  const fetchDocument = async () => {
-    if (!documentId) return;
-    setLoading(true);
-    setError(null);
-    try {
-      const res = await api.documents.getById(documentId);
-      setDocument(res.data);
-    } catch (err: any) {
-      setError(err.response?.data?.error || err.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  if (!isOpen) return null;
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/50 backdrop-blur-sm">
-      <div className="bg-white rounded-xl shadow-2xl w-full max-w-4xl max-h-[90vh] mx-4 flex flex-col">
-        <div className="flex items-center justify-between px-6 py-4 border-b border-claude-border">
-          <h3 className="text-lg font-semibold text-claude-text truncate pr-4">
-            {document?.title || 'Документ'}
-          </h3>
-          <button onClick={onClose} className="p-2 hover:bg-gray-100 rounded-lg transition-colors flex-shrink-0">
-            <X size={20} className="text-claude-subtext" />
-          </button>
-        </div>
-        <div className="flex-1 overflow-auto p-6">
-          {loading ? (
-            <div className="flex items-center justify-center py-12">
-              <RefreshCw size={24} className="animate-spin text-claude-subtext mr-3" />
-              <span>Завантаження...</span>
-            </div>
-          ) : error ? (
-            <div className="text-center py-12">
-              <XCircle size={32} className="mx-auto mb-3 text-red-400" />
-              <p className="text-red-600">{error}</p>
-            </div>
-          ) : document ? (
-            <div className="space-y-4">
-              <div className="bg-gray-50 rounded-lg p-4 space-y-2 text-xs">
-                {document.case_number && <div><span className="text-claude-subtext">Номер справи:</span> <span className="font-mono">{document.case_number}</span></div>}
-                {document.court && <div><span className="text-claude-subtext">Суд:</span> <span>{document.court}</span></div>}
-                {document.date && <div><span className="text-claude-subtext">Дата:</span> <span>{formatDate(document.date)}</span></div>}
-                {document.dispute_category && <div><span className="text-claude-subtext">Категорія:</span> <span>{document.dispute_category}</span></div>}
-              </div>
-              {document.full_text ? (
-                <div className="whitespace-pre-wrap text-sm text-claude-text bg-gray-50 rounded-lg p-4 max-h-96 overflow-auto">
-                  {document.full_text}
-                </div>
-              ) : document.full_text_html ? (
-                <div className="text-sm text-claude-text bg-gray-50 rounded-lg p-4 max-h-96 overflow-auto"
-                  dangerouslySetInnerHTML={{ __html: document.full_text_html }}
-                />
-              ) : (
-                <div className="text-center py-8 text-claude-subtext">
-                  <FileText size={32} className="mx-auto mb-2 text-gray-300" />
-                  <p className="text-sm">Повний текст документа недоступний</p>
-                </div>
-              )}
-            </div>
-          ) : null}
-        </div>
-        <div className="px-6 py-4 border-t border-claude-border flex justify-end">
-          <button onClick={onClose} className="px-4 py-2 text-sm border border-claude-border rounded-lg hover:bg-gray-50">Закрити</button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function ImportSamplesSection() {
-  const [data, setData] = useState<ImportSamplesData | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [hours, setHours] = useState(24);
-  const [expandedSources, setExpandedSources] = useState<Set<string>>(new Set());
-  const [selectedDocumentId, setSelectedDocumentId] = useState<string | null>(null);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const fetchSamples = useCallback(async (h: number = hours) => {
-    setLoading(true);
-    setError(null);
-    try {
-      const res = await api.admin.getImportSamples(h, 5);
-      setData(res.data);
-    } catch (err: any) {
-      setError(err.response?.data?.error || err.message);
-    } finally {
-      setLoading(false);
-    }
-  }, [hours]);
-
-  useEffect(() => {
-    fetchSamples();
-  }, [fetchSamples]);
-
-  const toggleSource = (source: string) => {
-    const newSet = new Set(expandedSources);
-    if (newSet.has(source)) {
-      newSet.delete(source);
-    } else {
-      newSet.add(source);
-    }
-    setExpandedSources(newSet);
-  };
-
-  const openDocument = (documentId: string) => {
-    setSelectedDocumentId(documentId);
-    setIsModalOpen(true);
-  };
-
-  const closeModal = () => {
-    setIsModalOpen(false);
-    setSelectedDocumentId(null);
-  };
-
-  if (loading && !data) return <SectionLoader />;
-  if (error && !data) return <SectionError message={error} onRetry={() => fetchSamples()} />;
-
-  const hasAnyData = data && data.samples && data.samples.length > 0;
-
-  return (
-    <div>
-      {/* Controls */}
-      <div className="flex items-center gap-3 mb-4">
-        <select
-          value={hours}
-          onChange={(e) => {
-            const h = Number(e.target.value);
-            setHours(h);
-            fetchSamples(h);
-          }}
-          className="text-xs border border-claude-border rounded-lg px-2 py-1.5 bg-white text-claude-text"
-        >
-          <option value={1}>Остання година</option>
-          <option value={6}>Останні 6 годин</option>
-          <option value={12}>Останні 12 годин</option>
-          <option value={24}>Останні 24 години</option>
-          <option value={48}>Останні 2 дні</option>
-          <option value={72}>Останні 3 дні</option>
-          <option value={168}>Останній тиждень</option>
-        </select>
-        <button
-          onClick={() => fetchSamples()}
-          disabled={loading}
-          className="flex items-center gap-1.5 px-3 py-1.5 text-xs bg-white border border-claude-border rounded-lg hover:bg-claude-bg transition-colors disabled:opacity-50"
-        >
-          <RefreshCw size={12} className={loading ? 'animate-spin' : ''} />
-          Оновити
-        </button>
-        {data && (
-          <span className="text-xs text-claude-subtext">
-            Знайдено {data.samples.length} джерел
-          </span>
-        )}
-      </div>
-
-      {/* Summary cards */}
-      {data && (
-        <div className="grid grid-cols-2 md:grid-cols-4 gap-3 mb-4">
-          <div className="bg-white rounded-lg border border-claude-border p-3">
-            <div className="text-lg font-semibold text-claude-text font-mono">{formatNumber(data.summary.court_decisions)}</div>
-            <div className="text-[10px] text-claude-subtext">Судові рішення</div>
-          </div>
-          <div className="bg-white rounded-lg border border-claude-border p-3">
-            <div className="text-lg font-semibold text-claude-text font-mono">{formatNumber(data.summary.legislation)}</div>
-            <div className="text-[10px] text-claude-subtext">Законодавство</div>
-          </div>
-          <div className="bg-white rounded-lg border border-claude-border p-3">
-            <div className="text-lg font-semibold text-claude-text font-mono">{formatNumber(data.summary.embeddings)}</div>
-            <div className="text-[10px] text-claude-subtext">Вектори</div>
-          </div>
-          <div className="bg-white rounded-lg border border-claude-border p-3">
-            <div className="text-lg font-semibold text-claude-text font-mono">{formatNumber(data.summary.user_uploads)}</div>
-            <div className="text-[10px] text-claude-subtext">Завантаження</div>
-          </div>
-        </div>
-      )}
-
-      {/* Samples by source */}
-      {hasAnyData ? (
-        <div className="space-y-3">
-          {data?.samples.map((sample) => (
-            <div key={sample.source} className="bg-white rounded-xl border border-claude-border overflow-hidden">
-              {/* Header */}
-              <button
-                onClick={() => toggleSource(sample.source)}
-                className="w-full flex items-center justify-between px-4 py-3 hover:bg-gray-50/50 transition-colors"
-              >
-                <div className="flex items-center gap-3">
-                  {expandedSources.has(sample.source) ? (
-                    <ChevronUp size={14} className="text-claude-subtext" />
-                  ) : (
-                    <ChevronDown size={14} className="text-claude-subtext" />
-                  )}
-                  <span className="font-medium text-sm text-claude-text">{sample.source_name}</span>
-                  <span className="inline-flex items-center px-2 py-0.5 rounded-full text-[10px] font-medium bg-blue-50 text-blue-700">
-                    {formatNumber(sample.count)}
-                  </span>
-                </div>
-                <span className="text-[10px] text-claude-subtext">
-                  {formatDate(sample.last_import)}
-                </span>
-              </button>
-
-              {/* Records */}
-              {expandedSources.has(sample.source) && (
-                <div className="border-t border-claude-border/30">
-                  {sample.records.map((record, idx) => (
-                    <div key={`${record.id}-${idx}`} className="px-4 py-2 border-b border-claude-border/20 last:border-b-0 hover:bg-gray-50/30">
-                      <div className="flex items-start justify-between gap-4">
-                        <div className="flex-1 min-w-0">
-                          {record.title && (
-                            <div className="text-xs text-claude-text truncate cursor-pointer hover:text-blue-600" title={record.title} onClick={() => openDocument(record.id)}>
-                              {record.title}
-                            </div>
-                          )}
-                          {record.document_title && (
-                            <div className="text-xs text-claude-text truncate" title={record.document_title}>
-                              {record.document_title}
-                            </div>
-                          )}
-                          {record.name && (
-                            <div className="text-xs text-claude-text truncate">
-                              {record.name}
-                            </div>
-                          )}
-                          <div className="flex items-center gap-2 mt-1 text-[10px] text-claude-subtext">
-                            {record.court && <span>{record.court}</span>}
-                            {record.case_number && <span className="font-mono">{record.case_number}</span>}
-                            {record.category && <span className="text-blue-600">{record.category}</span>}
-                            {record.justice_kind && <span className="text-purple-600">Вид: {record.justice_kind}</span>}
-                            {record.type && <span className="text-green-600">{record.type}</span>}
-                            {record.rada_id && <span className="font-mono">ID: {record.rada_id}</span>}
-                            {record.status && <span className="text-orange-600">{record.status}</span>}
-                            {record.vector_id && <span className="font-mono text-xs">{record.vector_id.substring(0, 20)}...</span>}
-                            {record.document_section_id && <span className="text-gray-500">Секція</span>}
-                            {record.user_email && <span>{record.user_email}</span>}
-                            {record.user_name && <span>{record.user_name}</span>}
-                            {record.domain && <span className="text-gray-500">{record.domain}</span>}
-                            <span className="text-gray-400">{formatDate(record.created_at)}</span>
-                          </div>
-                        </div>
-                        <button
-                          onClick={() => openDocument(record.id)}
-                          className="flex-shrink-0 p-1.5 text-claude-subtext hover:text-blue-600 hover:bg-blue-50 rounded-lg transition-colors"
-                          title="Переглянути документ"
-                        >
-                          <Eye size={14} />
-                        </button>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              )}
-            </div>
-          ))}
-        </div>
-      ) : data ? (
-        <div className="bg-white rounded-xl border border-claude-border p-8 text-center">
-          <Database size={24} className="mx-auto mb-2 text-claude-subtext" />
-          <p className="text-sm text-claude-subtext">Немає нових даних за обраний період</p>
-          <p className="text-[10px] text-claude-subtext/70 mt-1">
-            Скрипти імпорту не запускалися або не додали нових записів за {hours} год
-          </p>
-        </div>
-      ) : null}
-
-      {/* Document Modal */}
-      <DocumentModal
-        documentId={selectedDocumentId}
-        isOpen={isModalOpen}
-        onClose={closeModal}
-      />
-    </div>
-  );
-}
-
-export function AdminMonitoringPage() {
-  const [backend, setBackend] = useState<SectionState<BackendData>>({ data: null, loading: true, error: null });
-  const [rada, setRada] = useState<SectionState<ServiceData>>({ data: null, loading: true, error: null });
-  const [openreyestr, setOpenreyestr] = useState<SectionState<ServiceData>>({ data: null, loading: true, error: null });
-  const [courtDocs, setCourtDocs] = useState<SectionState<CourtDocsData>>({ data: null, loading: true, error: null });
-  const [courtDocsDays, setCourtDocsDays] = useState(30);
-
-  const fetchBackend = useCallback(async () => {
-    setBackend(prev => ({ ...prev, loading: true, error: null }));
-    try {
-      const res = await api.admin.getDataSources('backend');
-      setBackend({ data: res.data, loading: false, error: null });
-    } catch (err: any) {
-      setBackend(prev => ({ ...prev, loading: false, error: err.response?.data?.error || err.message }));
-    }
-  }, []);
-
-  const fetchRada = useCallback(async () => {
-    setRada(prev => ({ ...prev, loading: true, error: null }));
-    try {
-      const res = await api.admin.getDataSources('rada');
-      const d = res.data;
-      if (d.error && Object.keys(d.tables || {}).length === 0) {
-        setRada({ data: null, loading: false, error: d.error });
-      } else {
-        setRada({ data: d, loading: false, error: null });
-      }
-    } catch (err: any) {
-      setRada(prev => ({ ...prev, loading: false, error: err.response?.data?.error || err.message }));
-    }
-  }, []);
-
-  const fetchOpenreyestr = useCallback(async () => {
-    setOpenreyestr(prev => ({ ...prev, loading: true, error: null }));
-    try {
-      const res = await api.admin.getDataSources('openreyestr');
-      const d = res.data;
-      if (d.error && Object.keys(d.tables || {}).length === 0) {
-        setOpenreyestr({ data: null, loading: false, error: d.error });
-      } else {
-        setOpenreyestr({ data: d, loading: false, error: null });
-      }
-    } catch (err: any) {
-      setOpenreyestr(prev => ({ ...prev, loading: false, error: err.response?.data?.error || err.message }));
-    }
-  }, []);
-
-  const fetchCourtDocs = useCallback(async (days?: number) => {
-    const d = days ?? courtDocsDays;
-    setCourtDocs(prev => ({ ...prev, loading: true, error: null }));
-    try {
-      const res = await api.admin.getRecentCourtDocs(d, 5);
-      setCourtDocs({ data: res.data, loading: false, error: null });
-    } catch (err: any) {
-      setCourtDocs(prev => ({ ...prev, loading: false, error: err.response?.data?.error || err.message }));
-    }
-  }, [courtDocsDays]);
-
-  const fetchAll = useCallback(() => {
-    fetchBackend();
-    fetchRada();
-    fetchOpenreyestr();
-    fetchCourtDocs();
-  }, [fetchBackend, fetchRada, fetchOpenreyestr, fetchCourtDocs]);
-
-  useEffect(() => {
-    fetchAll();
-  }, [fetchAll]);
-
-  const anyLoading = backend.loading || rada.loading || openreyestr.loading;
-
-  const backendRows = backend.data?.tables?.reduce((s, t) => s + t.rows, 0) || 0;
-  const radaTables = rada.data ? toTableInfoArray(rada.data.tables) : [];
-  const radaRows = radaTables.reduce((s, t) => s + t.rows, 0);
-  const orTables = openreyestr.data ? toTableInfoArray(openreyestr.data.tables) : [];
-  const orRows = orTables.reduce((s, t) => s + t.rows, 0);
-  const totalRows = backendRows + radaRows + orRows;
-  const totalTables = (backend.data?.tables?.length || 0) + radaTables.length + orTables.length;
-
-  return (
-    <div className="flex-1 overflow-y-auto p-6">
-      {/* Header */}
-      <div className="flex items-center justify-between mb-6">
-        <div>
-          <h1 className="text-2xl font-semibold text-claude-text font-sans">Моніторинг джерел даних</h1>
-        </div>
-        <button
-          onClick={fetchAll}
-          disabled={anyLoading}
-          className="flex items-center gap-2 px-4 py-2 bg-white border border-claude-border rounded-lg text-sm text-claude-text hover:bg-claude-bg transition-colors disabled:opacity-50"
-        >
-          <RefreshCw size={14} className={anyLoading ? 'animate-spin' : ''} />
-          Оновити
-        </button>
-      </div>
-
-      {/* Summary Cards */}
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4 mb-8">
-        <SummaryCard
-          icon={Layers}
-          label="Всього записів"
-          value={formatNumber(totalRows)}
-          sub={`${totalTables} таблиць`}
-          loading={anyLoading && totalRows === 0}
-        />
-        <SummaryCard
-          icon={HardDrive}
-          label="Backend DB"
-          value={`${backend.data?.dbSizeMb || 0} MB`}
-          sub={`${formatNumber(backendRows)} записів`}
-          status={backend.loading ? 'loading' : 'online'}
-        />
-        <SummaryCard
-          icon={Server}
-          label="RADA DB"
-          value={rada.data ? `${rada.data.dbSizeMb || 0} MB` : '—'}
-          sub={rada.loading ? 'Завантаження...' : rada.data ? `${formatNumber(radaRows)} записів` : 'Недоступний'}
-          status={rada.loading ? 'loading' : rada.data ? 'online' : 'offline'}
-        />
-        <SummaryCard
-          icon={Database}
-          label="OpenReyestr DB"
-          value={openreyestr.data ? `${openreyestr.data.dbSizeMb || 0} MB` : '—'}
-          sub={openreyestr.loading ? 'Завантаження...' : openreyestr.data ? `${formatNumber(orRows)} записів` : 'Недоступний'}
-          status={openreyestr.loading ? 'loading' : openreyestr.data ? 'online' : 'offline'}
-        />
-      </div>
-
-      {/* Backend Sources */}
-      <section className="mb-8">
-        <div className="flex items-center justify-between mb-4">
-          <div className="flex items-center gap-2">
-            <Activity size={18} className="text-claude-subtext" />
-            <h2 className="text-lg font-semibold text-claude-text font-sans">Backend (mcp_backend)</h2>
-            {!backend.loading && <ServiceStatusBadge available={!backend.error} />}
-          </div>
-          {backend.data && (
-            <span className="text-xs text-claude-subtext bg-claude-bg px-2 py-1 rounded-full">
-              PostgreSQL :5432 · {backend.data.dbSizeMb} MB
-            </span>
-          )}
-        </div>
-        {backend.loading ? (
-          <SectionLoader />
-        ) : backend.error ? (
-          <SectionError message={backend.error} onRetry={fetchBackend} />
-        ) : backend.data ? (
-          <DataTable tables={backend.data.tables} />
-        ) : null}
-      </section>
-
-      {/* Court Documents by Practice Area */}
-      <section className="mb-8">
-        <div className="flex items-center justify-between mb-4">
-          <div className="flex items-center gap-2">
-            <Scale size={18} className="text-claude-subtext" />
-            <h2 className="text-lg font-semibold text-claude-text font-sans">Судові рішення за видами права</h2>
-          </div>
-          <div className="flex items-center gap-2">
-            <select
-              value={courtDocsDays}
-              onChange={(e) => {
-                const d = Number(e.target.value);
-                setCourtDocsDays(d);
-                fetchCourtDocs(d);
-              }}
-              className="text-xs border border-claude-border rounded-lg px-2 py-1.5 bg-white text-claude-text"
-            >
-              <option value={7}>7 днів</option>
-              <option value={30}>30 днів</option>
-              <option value={90}>90 днів</option>
-              <option value={365}>1 рік</option>
-            </select>
-          </div>
-        </div>
-        <CourtDocsSection state={courtDocs} onRetry={() => fetchCourtDocs()} />
-      </section>
-
-      {/* Document Completeness Check */}
-      <section className="mb-8">
-        <div className="flex items-center gap-2 mb-4">
-          <CheckCircle size={18} className="text-claude-subtext" />
-          <h2 className="text-lg font-semibold text-claude-text font-sans">Перевірка повноти документів</h2>
-        </div>
-        <DocumentCompletenessSection />
-      </section>
-
-      {/* Court Registry Scraper */}
-      <section className="mb-8">
-        <div className="flex items-center gap-2 mb-4">
-          <Download size={18} className="text-claude-subtext" />
-          <h2 className="text-lg font-semibold text-claude-text font-sans">Докачати документи з реєстру</h2>
-        </div>
-        <CourtDataMapSection />
-      </section>
-
-      {/* Import Samples - Recent Script Uploads */}
-      <section className="mb-8">
-        <div className="flex items-center gap-2 mb-4">
-          <Database size={18} className="text-claude-subtext" />
-          <h2 className="text-lg font-semibold text-claude-text font-sans">Зразки даних з останніх завантажень</h2>
-        </div>
-        <ImportSamplesSection />
-      </section>
-
-      {/* RADA Sources */}
-      <section className="mb-8">
-        <div className="flex items-center justify-between mb-4">
-          <div className="flex items-center gap-2">
-            <Server size={18} className="text-claude-subtext" />
-            <h2 className="text-lg font-semibold text-claude-text font-sans">RADA Server (mcp_rada)</h2>
-            {!rada.loading && <ServiceStatusBadge available={!!rada.data} />}
-          </div>
-          {rada.data && (
-            <span className="text-xs text-claude-subtext bg-claude-bg px-2 py-1 rounded-full">
-              PostgreSQL :5433 · {rada.data.dbSizeMb} MB
-            </span>
-          )}
-        </div>
-        {rada.loading ? (
-          <SectionLoader />
-        ) : rada.error ? (
-          <SectionError message={rada.error} onRetry={fetchRada} />
-        ) : radaTables.length > 0 ? (
-          <DataTable tables={radaTables} />
-        ) : (
-          <div className="bg-white rounded-xl border border-claude-border p-8 text-center">
-            <AlertTriangle size={24} className="mx-auto mb-2 text-yellow-500" />
-            <p className="text-sm text-claude-subtext">RADA сервер не повертає статистику</p>
-          </div>
-        )}
-      </section>
-
-      {/* OpenReyestr Sources */}
-      <section className="mb-8">
-        <div className="flex items-center justify-between mb-4">
-          <div className="flex items-center gap-2">
-            <Database size={18} className="text-claude-subtext" />
-            <h2 className="text-lg font-semibold text-claude-text font-sans">OpenReyestr Server (mcp_openreyestr)</h2>
-            {!openreyestr.loading && <ServiceStatusBadge available={!!openreyestr.data} />}
-          </div>
-          {openreyestr.data && (
-            <span className="text-xs text-claude-subtext bg-claude-bg px-2 py-1 rounded-full">
-              PostgreSQL :5435 · {openreyestr.data.dbSizeMb} MB
-            </span>
-          )}
-        </div>
-        {openreyestr.loading ? (
-          <SectionLoader />
-        ) : openreyestr.error ? (
-          <SectionError message={openreyestr.error} onRetry={fetchOpenreyestr} />
-        ) : orTables.length > 0 ? (
-          <DataTable tables={orTables} />
-        ) : (
-          <div className="bg-white rounded-xl border border-claude-border p-8 text-center">
-            <AlertTriangle size={24} className="mx-auto mb-2 text-yellow-500" />
-            <p className="text-sm text-claude-subtext">OpenReyestr сервер не повертає статистику</p>
-          </div>
-        )}
-      </section>
-
-      {/* Recent Imports (OpenReyestr) */}
-      {openreyestr.data?.recentImports && openreyestr.data.recentImports.length > 0 && (
-        <section className="mb-8">
-          <div className="flex items-center gap-2 mb-4">
-            <Clock size={18} className="text-claude-subtext" />
-            <h2 className="text-lg font-semibold text-claude-text font-sans">Останні імпорти (OpenReyestr)</h2>
-          </div>
-          <div className="bg-white rounded-xl border border-claude-border shadow-sm overflow-hidden">
-            <div className="overflow-x-auto">
-              <table className="w-full text-sm">
-                <thead>
-                  <tr className="border-b border-claude-border bg-gray-50">
-                    <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Реєстр</th>
-                    <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Статус</th>
-                    <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Імпортовано</th>
-                    <th className="text-right px-4 py-2.5 font-medium text-claude-subtext text-xs">Помилок</th>
-                    <th className="text-left px-4 py-2.5 font-medium text-claude-subtext text-xs">Завершено</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {openreyestr.data.recentImports.map((imp: any, i: number) => (
-                    <tr key={i} className="border-b border-claude-border/30 hover:bg-gray-50/50">
-                      <td className="px-4 py-2.5 text-xs font-medium text-claude-text">{imp.registry_name}</td>
-                      <td className="px-4 py-2.5">
-                        <span className={`inline-block px-2 py-0.5 rounded-full text-[10px] font-medium ${
-                          imp.status === 'completed' ? 'bg-green-50 text-green-700' :
-                          imp.status === 'failed' ? 'bg-red-50 text-red-700' :
-                          'bg-yellow-50 text-yellow-700'
-                        }`}>
-                          {imp.status}
-                        </span>
-                      </td>
-                      <td className="px-4 py-2.5 text-right text-xs font-mono">{formatNumber(imp.records_imported)}</td>
-                      <td className="px-4 py-2.5 text-right text-xs font-mono">
-                        <span className={imp.records_failed > 0 ? 'text-red-600' : ''}>{formatNumber(imp.records_failed)}</span>
-                      </td>
-                      <td className="px-4 py-2.5 text-xs text-claude-subtext">{formatDate(imp.import_completed_at)}</td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          </div>
-        </section>
-      )}
-    </div>
-  );
-}
+export { AdminMonitoringPage } from './admin/monitoring/index';
