+++ lexwebapp/src/pages/admin/monitoring/constants.ts
@@ -1,0 +1,16 @@
+export const JUSTICE_KINDS = [
+  { label: 'Кримінальне', value: 'Кримінальне', id: '5' },
+  { label: 'Цивільне', value: 'Цивільне', id: '1' },
+  { label: 'Адміністративне', value: 'Адміністративне', id: '2' },
+  { label: 'Господарське', value: 'Господарське', id: '3' },
+];
+
+export const DOC_FORMS = [
+  { label: 'Усі форми', value: '__all__' },
+  { label: 'Рішення', value: 'Рішення' },
+  { label: 'Ухвала', value: 'Ухвала' },
+  { label: 'Вирок', value: 'Вирок' },
+  { label: 'Постанова', value: 'Постанова' },
+];
+
+export const MONTHS_SHORT = ['С','Л','Б','К','Т','Ч','Л','С','В','Ж','Л','Г'];
